fix(1102): warm runner pipeline in e2e beforeAll to kill headline flake
The auth-error-recovery headline test regressed under full-suite load after
#1115 because the e2e file got only half of that fix: CredentialStore was
hoisted to beforeAll (paying PBKDF2 100k cost once), but the SpellCaster
pipeline warm-up was not. The first `caster.run(...)` in a fresh worker
pays a ~500 ms cold-start tax — platform-sandbox detection
(`_cachedDetection` in platform-sandbox.ts), validator, prereq-checker,
step-executor — which under fork contention tipped the headline test past
the 5 s per-test ceiling.

Extract the warm-up into a shared helper (`warmRunnerPipeline`) so both
auth-error-runner.test.ts and auth-error-recovery-e2e.test.ts pay that tax
inside beforeAll's 10 s ceiling. Replaces the inline warm-up in
auth-error-runner.test.ts (DRY) with the shared helper.

Verification: 6 consecutive full-suite npm test runs — headline test green
in all 6 (#1102 fix holds). Touched files green 14/14 in isolation (2.35 s).
